Tweak CounterTable so the interface doesn't need net-snmp libraries included
